Do not try to coerce "" for numeric & date params
When using an HTML form for searching or filtering and a field is left
blank, Rails puts an empty string in the params hash, and this raised
an unexpected InvalidParameterError when we defined the param as a
non-required Integer, Float, Date, etc.

Now we will treat that empty string as an un-supplied param and only
reject it if the param is flagged as required.

- added shared examples for Float tests
- nested shared examples in the appropriate scope